fix(web): fix line breaks in currency values
